# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Firestick API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/graph/basic": {
      "get": {
        "summary": "Basic graph",
        "parameters": [ { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1 } } ],
        "responses": { "200": { "description": "Graph data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GraphResponse" } } } }, "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }, "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
    "/api/graph/enriched": {
      "get": {
        "summary": "Enriched graph",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "include", "in": "query", "schema": { "type": "string" }, "description": "comma list: containment,imports,calls,inheritance" },
          { "name": "nodeTypes", "in": "query", "schema": { "type": "string" }, "description": "comma list: F,C,M,I" }
        ],
        "responses": { "200": { "description": "Graph data", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GraphResponse" } } } }, "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }, "500": { "description": "Server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "timestamp": { "type": "string", "format": "date-time" },
          "path": { "type": "string" },
          "requestId": { "type": "string", "description": "Correlation id; also returned in X-Request-Id header" },
          "status": { "type": "integer" },
          "error": { "type": "string" },
          "code": { "type": "string" },
          "message": { "type": "string" },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
...
      "GraphResponse": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GraphNode" }
          },
          "edges": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/GraphEdge" }
          },
          "metadata": { "type": "object", "additionalProperties": true }
        }
      },
      "GraphNode": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "label": { "type": "string" },
          "type": { "type": "string", "enum": ["F", "C", "M", "I"] }
        }
      },
      "GraphEdge": {
        "type": "object",
        "properties": {
          "source": { "type": "string" },
          "target": { "type": "string" },
          "type": { "type": "string" }
        }
      },
...
    }
  }
}